requirements.txt
python-cfworker
requests==2.5.1
Flask
textblob
simplejson
psutil
